Improve bot greatly
Ad farming has been notoriously hard since the X to close the ad is
pretty much random and going back to Battle Cats would crash the ad and
not give a reward.

Now, if you simply turn off the navigation bar with a gesture bar
instead, the X stays in a constant position and farming is much more
reliable. Extra code has been added to catch the cases were the ad is
opened instead.