Try #1 in fixing the cursed "midnight bug"
The midnight bug is a bug where the bot gets completely stuck at
midnight every day. This occurs because the daily events, rewards, and
deals all pop up at this time with wacky buttons to close them. This has
been tried and proven to successfully close the daily rewards and events
screen, but not the deal screen.